+++ packages/utils/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## [1.4.0] - 2026-09-01
 
 ### New Features